Make the toad rest between jumps
We don't want the toad to "continuously" jump when the direction buttons are held
since that's not faithful to the original game and it's not as fun.

Instead we want each jump to be a discrete movement even if the player is holding
a direction key.

To achieve this we can force a delay between finishing one jump and starting
another jump. Which is what this commit does