Remove Physics.Body
Instead of using `Body` we can use `HasPosition` and `HasJumpMotion` to determine
which entities are capable of supporting an action.

For example: `JumpMotion` updates the position of an entity. So we just
need to be passed an entity that has a `JumpMotion` and a `Position` which
we can use classy lenses for.