Make the Rendering effect command-based
Currently we can't test anything to do with rendering. The plan is to use
this command structure to let us test things that are doing rendering.

This will become more important as we begin to build animation. But it will also
let us fill out our test suite, which has been a bit neglected in the rendering
department.